+++ docs/tutorials/pisa/pisa.Person-1.pdf
@@ -70,7 +70,7 @@ endobj
 % 'R7': class PDFInfo 
 7 0 obj
 << /Author ()
- /CreationDate (D:20150810060821-02'00')
+ /CreationDate (D:20150810165406-02'00')
  /Creator (\(unspecified\))
  /Keywords ()
  /Producer (pisa HTML to PDF <http://www.htmltopdf.org>)
@@ -124,7 +124,7 @@ xref
 trailer
 << /ID 
  % ReportLab generated PDF document -- digest (http://www.reportlab.com) 
- [(\001^D\311\273Vv\302\321de\322Y\245<\337) (\001^D\311\273Vv\302\321de\322Y\245<\337)] 
+ [(C\211\204R\342\177\312\010g\003\(q\026=\373\354) (C\211\204R\342\177\312\010g\003\(q\026=\373\354)] 
 
  /Info 7 0 R
  /Root 6 0 R
