+++ lino/modlib/print_pisa/models.py
@@ -44,7 +44,8 @@ class PrintTableActionPisa(PrintTableAction):
 
         extend_context(context)
 
-        template = settings.SITE.jinja_env.get_template(self.template_name)
+        template = settings.SITE.plugins.jinja.renderer.jinja_env.get_template(
+            self.template_name)
         html = template.render(**context).encode('utf-8')
 
         with open(output_file + '.html', "w") as file:
